Demo Completion

+++ src/app/pages/header/header.component.ts
@@ -10,9 +10,11 @@ import { LogService } from '../../services/logs/log.service';
 export class HeaderComponent implements OnInit {
 
   role: string;
+  logId: string;
 
   constructor(public logServices: LogService) { 
     this.role = sessionStorage.getItem('role');
+    this.logId = sessionStorage.getItem('sessionId');
   }
 
   ngOnInit(): void {
@@ -20,30 +22,33 @@ export class HeaderComponent implements OnInit {
 
   signOut() {
 
-    sessionStorage.clear();
+    if(this.role == "0") {
 
-    let data = {
-      logId: ""
+      let data = {
+        logId: this.logId
+      }
+  
+      this.logServices.update(data).subscribe(
+        (res: any) => {
+
+          if(res.status === 200) {
+            sessionStorage.clear();
+            location.href = "/authentication";
+          } else {
+            alert("Oops! Problem Occured, Please Try Again Later.");
+          }
+          // console.log(res);
+  
+        }, (error) => {
+            alert("Oops! Problem Occured, Please Try Again Later.");
+            // console.log(error);
+        }
+      );
+    } else {
+      sessionStorage.clear();
+      location.href = "/authentication";
     }
-
-    location.href = "/authentication";
-
-    console.log("logout");
     
-    // this.logServices.update(data).subscribe(
-    //   (res: any) => {
-    //     if(res.status === 200) {
-    //       location.href="/";
-    //     } else {
-    //       alert("Oops! Problem Occured, Please Try Again Later.");
-    //     }
-    //     // console.log(res);
-
-    //   }, (error) => {
-    //       alert("Oops! Problem Occured, Please Try Again Later.");
-    //       console.log(error);
-    //   }
-    // );
 
   }
